Refactor cql-utils and add adapterconfgen
